Working on broker container, subscriber template
Now the repository contains two Dockerfiles. The Dockerfile_broker,
together with the entrypoint file, allow the configuration of a custom
mosquitto image and the Dockerfile_subscriber allows the configuration
of a custom python app that subscribes to topics.

The broker folder contains broker configurations file, the mqtt_app
contains the subscriber config files including the source code.