Updating certificate creation script
Now a certificate authority is created and a certificate signing request
is issued to the CA. Everything is done from the proxy/gen-certs script
and every file is generated inside the proxy/certificates folder.
The certificates folder is copied inside the container at build time but
can also be mapped from the docker compose file.

However, the certificate won't be trusted by browsers because the CA is
not recognized. It must be trusted by the user with specific consent.